adding redundant parts of the POM for 'mediation screens'
