+++ playwright-e2e/pages/exui/solicitor-events/response/mediation/mediation-availability/mediation-availability-content.ts
@@ -1,10 +1,4 @@
-import { date } from '../../../../../../../e2e/api/dataHelper';
 import { Party } from '../../../../../../models/partys';
-
-export const heading = {
-  claimant: { label: 'View and respond to defence' },
-  defendant: { label: 'Respond to claim' },
-};
 
 export const subHeading = 'Mediation availability';
